requirements.txt
langchain==0.3.14
langchain-chroma==0.1.4
langchain-community==0.3.14
langchain-core==0.3.29
langchain-openai==0.2.14
langchain-text-splitters==0.3.5
langdetect==1.0.9
langsmith==0.2.10
openai==1.59.4
python-dotenv==1.0.1
python-pptx==1.0.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0 
tomli==2.2.1
fastapi==0.115.6
fastapi-utils==0.8.0
uvicorn==0.34.0
uvloop==0.21.0
typing-inspect==0.9.0
pypdf==5.1.0
unstructured==0.16.12
unstructured-client==0.28.1
docx2txt==0.8